+++ js/elements.js
@@ -1,23 +1,43 @@
 /*
  * Document Query Selector(s)
  */
-function $(elms, callback) {
+function $(elms, evt, callback) {
     var gElms = document.querySelectorAll(elms);
     var gElms_length = gElms.length;
+
     if (gElms_length > 0) {
         if (gElms_length > 1) {
+
             if (typeof callback === 'function') {
                 for (var i = 0; i < gElms_length; i++) {
-                    callback(gElms[i]);
+
+                    if (evt === null) {
+                        callback(gElms[i]);
+                    } else {
+                        on(evt, gElms[i], function (elm, e) {
+                            callback(elm, e);
+                        });
+                    }
                 }
+
             } else {
                 return gElms;
             }
+
         } else {
+            var delms = document.querySelector(elms);
             if (typeof callback === 'function') {
-                callback(document.querySelector(elms));
+
+                if (evt === null) {
+                    callback(delms);
+                } else {
+                    on(evt, delms, function (elm, e) {
+                        callback(elm, e);
+                    });
+                }
+
             } else {
-                return document.querySelector(elms);
+                return delms;
             }
         }
     } else {